add probabilty functions to new file
