+++ velox/docs/functions/presto/math.rst
@@ -6,12 +6,6 @@ Mathematical Functions
 
     Returns the absolute value of ``x``.
 
-.. function:: beta_cdf(a, b, value) -> double
-
-    Compute the Beta cdf with given a, b parameters:  P(N < value; a, b).
-    The a, b parameters must be positive real numbers and value must be a real value (all of type DOUBLE).
-    The value must lie on the interval [0, 1].
-
 .. function:: cbrt(x) -> double
 
     Returns the cube root of ``x``.
@@ -239,3 +233,14 @@ Floating Point Functions
 .. function:: nan() -> double
 
     Returns the constant representing not-a-number.
+
+
+====================================
+Probability Functions
+====================================
+
+.. function:: beta_cdf(a, b, value) -> double
+
+    Compute the Beta cdf with given a, b parameters:  P(N < value; a, b).
+    The a, b parameters must be positive real numbers and value must be a real value (all of type DOUBLE).
+    The value must lie on the interval [0, 1].
